requirements.txt
asgiref==3.5.2
Django==4.0.6
django-extensions==3.2.0
django-phonenumber-field==6.3.0
django-ratings==0.3.7
Pillow==9.2.0
python-decouple==3.6
reportlab==3.6.11
sqlparse==0.4.2
tzdata==2022.1
